Better organize KILT module
